pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mock-trader"
version = "0.1.0"
description = "Live mock trading system for ML/AI models on BTC/USDT"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9",
    "pandas>=2.0",
    "numpy>=1.24",
    "xgboost>=2.0",
    "joblib>=1.3",
    "requests>=2.28",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["UP017", "E741", "N806"]  # UP017: datetime.UTC; E741/N806: OHLCV l, ML X

[tool.ruff.lint.isort]
known-first-party = ["config", "db", "fetcher", "features", "trader", "monitor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
